feat: ограничить права пользователей в удалении чужих карточек

+++ controllers/cards.js
@@ -1,5 +1,7 @@
 const Card = require('../models/card');
-const { handleBadRequestError, handleDefaultError, handleNotFoundError } = require('../utils/errors');
+const {
+  handleBadRequestError, handleDefaultError, handleNotFoundError, handleForbiddenError,
+} = require('../utils/errors');
 
 function getCardsList(req, res) {
   Card.find({})
@@ -28,20 +30,40 @@ function createCard(req, res) {
 }
 
 function deleteCard(req, res) {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .orFail(() => {
       const error = new Error();
       error.name = 'DocumentNotFoundError';
       throw error;
     })
     .then((card) => {
-      res.send(card);
+      if (card.owner.toString() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((removedCard) => {
+            res.send(removedCard);
+          })
+          .catch((err) => {
+            if (err.name === 'CastError') {
+              handleBadRequestError(res);
+            } else if (err.name === 'DocumentNotFoundError') {
+              handleNotFoundError(res);
+            } else {
+              handleDefaultError(res);
+            }
+          });
+      } else {
+        const error = new Error();
+        error.name = 'ForbiddenError';
+        throw error;
+      }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
         handleBadRequestError(res);
       } else if (err.name === 'DocumentNotFoundError') {
         handleNotFoundError(res);
+      } else if (err.name === 'ForbiddenError') {
+        handleForbiddenError(res);
       } else {
         handleDefaultError(res);
       }
